Move all logging stuff to utils.py
