+++ method_comparison/MetaMathQA/run.py
@@ -18,21 +18,16 @@ TODO
 
 import argparse
 import datetime as dt
-import enum
 import gc
 import json
 import os
 import sys
-import tempfile
 import textwrap
 import time
 from contextlib import nullcontext
-from dataclasses import asdict, dataclass
 from functools import partial
 from typing import Any, Literal, Optional
 
-import huggingface_hub
-import numpy as np
 import torch
 from torch import nn
 from torch.cuda.amp import autocast, GradScaler
@@ -40,21 +35,21 @@ from tqdm import tqdm
 from transformers import set_seed, get_cosine_schedule_with_warmup
 
 from peft import PeftConfig
-from peft.utils import SAFETENSORS_WEIGHTS_NAME
 
 from data import get_dataset, get_filtered_dataset, get_train_valid_test_datasets, tokenize_with_answer, tokenize_wo_answer
 from utils import (
     DATASET_NAME,
     FILE_NAME_TRAIN_PARAMS,
+    TrainResult,
+    TrainStatus,
     get_accuracy,
     get_base_model_info,
     get_dataset_info,
-    get_meta_info,
     get_model,
-    get_peft_branch,
     get_tokenizer,
     get_train_config,
     init_cuda,
+    log_results,
     validate_experiment_path,
 )
 
@@ -63,33 +58,12 @@ from utils import (
 # warnings.filterwarnings("ignore") # FIXME?
 
 dtype_to_bytes_linear = {"float32": 4, "float16": 2, "bfloat16": 2, "int8": 1, "int4": 0.5}
-# main results
-RESULT_PATH = os.path.join(os.path.dirname(__file__), "results")
-# testing results
-RESULT_PATH_TEST = os.path.join(os.path.dirname(__file__), "temporary_results")
-# cancelled results
-RESULT_PATH_CANCELLED = os.path.join(os.path.dirname(__file__), "cancelled_results")
 # if lr scheduler with warmup is used, the ratio of warmup steps to total steps
 WARMUP_STEP_RATIO = 0.1
 
 # train/valid/test split
-VALID_SIZE = 40  # FIXME small, as it is performed every couple of steps
-TEST_SIZE = 1000  # FIXME can be larger
-
-
-class TrainStatus(enum.Enum):
-    FAILED = "failed"
-    SUCCESS = "success"
-    CANCELED = "canceled"
-
-
-@dataclass
-class TrainResult:
-    status: TrainStatus
-    train_time: float
-    cuda_memory_log: list[int]
-    losses: list[float]
-    metrics: list[Any]  # TODO
+VALID_SIZE = 50
+TEST_SIZE = 1000
 
 
 def evaluate(model, tokenizer, ds, batch_size, generate_kwargs, use_tqdm=False) -> tuple[list[str], list[str]]:
@@ -306,103 +280,6 @@ def train(
     return train_result
 
 
-def log_to_console(log_data: dict) -> None:
-    cuda_memory_max = log_data["train_info"]["cuda_memory_max"]
-    cuda_memory_avg = log_data["train_info"]["cuda_memory_avg"]
-    cuda_memory_99th = log_data["train_info"]["cuda_memory_99th"]
-    time_train = log_data["train_info"]["train_time"]
-    time_total = log_data["run_info"]["total_time"]
-    file_size = log_data["train_info"]["file_size"]
-
-    print(f"cuda memory max: {cuda_memory_max // 2**20}MB")
-    print(f"cuda memory avg: {cuda_memory_avg // 2**20}MB")
-    print(f"cuda memory 90th percentile: {cuda_memory_99th // 2**20}MB")
-    print(f"train time: {time_train}s")
-    print(f"total time: {time_total:.2f}s")
-    print(f"file size of checkpoint: {file_size / 2**20:.1f}MB")
-
-
-def log_to_file(*, log_data: dict, save_dir: str, experiment_name: str, timestamp: str) -> None:
-    file_name = f"{experiment_name.replace(os.path.sep, '--')}--{timestamp.replace(':', '-')}.json"
-    file_name = os.path.join(save_dir, file_name)
-    with open(file_name, "w") as f:
-        json.dump(log_data, f, indent=2)
-    print_verbose(f"Saved log to: {file_name}")
-
-
-def log_results(
-    *,
-    experiment_name: str,
-    train_result: TrainResult,
-    cuda_memory_init: int,
-    time_total: float,
-    model: nn.Module,
-    model_info: huggingface_hub.ModelInfo,
-    dataset_info: huggingface_hub.DatasetInfo,
-    start_date: str,
-    peft_config_sha: str,
-    train_params_sha: str,
-):
-    # collect results
-    cuda_memory_final = torch.cuda.max_memory_allocated()
-    cuda_memory_avg = int(sum(train_result.cuda_memory_log) / len(train_result.cuda_memory_log))
-    cuda_memory_99th = int(np.percentile(train_result.cuda_memory_log, 99))
-
-    with tempfile.TemporaryDirectory() as tmp_dir:
-        model.save_pretrained(tmp_dir)
-        stat = os.stat(os.path.join(tmp_dir, SAFETENSORS_WEIGHTS_NAME))
-        file_size = stat.st_size
-        print(f"Saved PEFT checkpoint to {tmp_dir}")
-
-    meta_info = get_meta_info()
-    model_sha = model_info.sha
-    model_created_at = model_info.created_at.isoformat()
-    dataset_sha = dataset_info.sha
-    dataset_created_at = dataset_info.created_at.isoformat()
-
-    peft_branch = get_peft_branch()
-
-    if train_result.status == TrainStatus.CANCELED:
-        save_dir = RESULT_PATH_CANCELLED
-        print_verbose("Experiment run was categorized as canceled")
-    elif peft_branch != "main":
-        save_dir = RESULT_PATH_TEST
-        print_verbose(f"Experiment run was categorized as a test run on branch {peft_branch}")
-    elif train_result.status == TrainStatus.SUCCESS:
-        save_dir = RESULT_PATH
-        print_verbose("Experiment run was categorized as successful run")
-
-    log_data = {
-        "run_info": {
-            "created_at": start_date,
-            "total_time": time_total,
-            "experiment_name": experiment_name,
-            "peft_branch": peft_branch,
-            "train_params_sha": train_params_sha,
-            "peft_config_sha": peft_config_sha,
-        },
-        "train_info": {
-            "cuda_memory_avg": cuda_memory_avg,
-            "cuda_memory_max": (cuda_memory_final - cuda_memory_init),
-            "cuda_memory_99th": cuda_memory_99th,
-            "train_time": train_result.train_time,
-            "file_size": file_size,
-            "status": train_result.status.value,
-            "metrics": train_result.metrics,
-        },
-        "meta_info": {
-            "model_sha": model_sha,
-            "model_created_at": model_created_at,
-            "dataset_sha": dataset_sha,
-            "dataset_created_at": dataset_created_at,
-            **asdict(meta_info),
-        },
-    }
-
-    log_to_console(log_data)
-    log_to_file(log_data=log_data, save_dir=save_dir, experiment_name=experiment_name, timestamp=start_date)
-
-
 def main(*, path_experiment: str, experiment_name: str, train_params_sha: str, peft_config_sha: str) -> None:
     tic_total = time.perf_counter()
     start_date = dt.datetime.now(tz=dt.timezone.utc).replace(microsecond=0).isoformat()
@@ -468,6 +345,7 @@ def main(*, path_experiment: str, experiment_name: str, train_params_sha: str, p
         start_date=start_date,
         train_params_sha=train_params_sha,
         peft_config_sha=peft_config_sha,
+        print_fn=print_verbose,
     )
